requirement.txt
fastapi
uvicorn
streamlit
pydantic
faiss-cpu
pypdf
python-docx
python-dotenv
numpy
requests
python-multipart
aiohttp
